pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpforge"
version = "0.1.0"
description = "Auto-generate Model Context Protocol (MCP) servers from API documentation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "model-context-protocol", "api", "openapi", "ai", "claude", "ollama"]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "mcp>=1.0.0",
]

[project.scripts]
mcpforge = "cli.main:app"

[project.urls]
Homepage = "https://acunningham-ship-it.github.io/mcpforge"
Repository = "https://github.com/acunningham-ship-it/mcpforge"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcpforge*", "cli*", "api*"]
